pyproject.toml
[tool.poetry]
name = "mailing"
version = "0.1.0"
description = ""
authors = ["Deadmau5"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
django-crontab = "^0.7.1"
ipython = "^8.14.0"
python-dotenv = "^1.0.0"
pillow = "^10.0.0"
redis = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
